requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
python-dotenv==1.0.1
langchain==0.2.11
langchain-community==0.2.10
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
python-box==7.2.0
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.3
PyPDF2==3.0.1
unstructured==0.15.0
tiktoken==0.7.0
pytest==8.3.2
